# docs/openapi/openapi.json
openapi: 3.0.3
info:
  title: StratTraderPro API
  version: 0.0.0
  description: Trading bot platform API — webhook-driven, regime-aware, multi-broker.
paths:
  /api/v1/auth/login/:
    post:
      operationId: auth_login
      description: Returns `{access, refresh, user, mfa_required}`. `mfa_required`
        is always false in M01 (MFA ships in M02).
      summary: Log in with email + password
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
            examples:
              LoginRequest:
                value:
                  email: trader@example.com
                  password: correct horse battery staple
                summary: Login request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Login'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Login'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                InvalidCredentials:
                  value:
                    error:
                      code: INVALID_CREDENTIALS
                      message: Invalid email or password.
                  summary: Invalid credentials
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                EmailNotVerified:
                  value:
                    error:
                      code: EMAIL_NOT_VERIFIED
                      message: Please verify your email before signing in.
                  summary: Email not verified
          description: ''
        '423':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                AccountLocked:
                  value:
                    error:
                      code: ACCOUNT_LOCKED
                      message: Account temporarily locked due to repeated failures.
                  summary: Account locked
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                RateLimited:
                  value:
                    error:
                      code: RATE_LIMITED
                      message: Too many requests.
                  summary: Rate limited
          description: ''
  /api/v1/auth/logout/:
    post:
      operationId: auth_logout
      summary: Log out — revokes refresh family
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Logout'
            examples:
              LogoutRequest:
                value:
                  refresh: eyJhbGciOi...refresh.jwt
                summary: Logout request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Logout'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Logout'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusEnvelope'
          description: ''
...
  /api/v1/auth/password/reset/:
    post:
      operationId: auth_password_reset
      description: Always returns 200 to avoid email enumeration.
      summary: Request a password reset email
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordReset'
            examples:
              ResetRequest:
                value:
                  email: trader@example.com
                summary: Reset request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordReset'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordReset'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                RateLimited:
                  value:
                    error:
                      code: RATE_LIMITED
                      message: Too many requests.
                  summary: Rate limited
          description: ''
  /api/v1/auth/password/reset/confirm/:
    post:
      operationId: auth_password_reset_confirm
      description: Issues a JWT pair on success and revokes any outstanding refresh
        families.
      summary: Set a new password using the reset token
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
            examples:
              ConfirmRequest:
                value:
                  token: prv_7a1c...sample
                  password: correct horse battery staple
                summary: Confirm request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirm'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairEnvelope'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                TokenInvalidOrExpired:
                  value:
                    error:
                      code: TOKEN_INVALID
                      message: Token is invalid or expired.
                  summary: Token invalid or expired
                PasswordDoesNotMeetPolicy:
                  value:
                    error:
                      code: PASSWORD_WEAK
                      message: Password does not meet policy.
                      details:
                        password:
                        - This password is too common.
                  summary: Password does not meet policy
          description: ''
  /api/v1/auth/refresh/:
    post:
      operationId: auth_refresh
      description: Consumes the supplied refresh token and issues a new pair in the
        same family. Reusing an already-rotated refresh revokes the whole family.
      summary: Rotate a refresh token
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Refresh'
            examples:
              RefreshRequest:
                value:
                  refresh: eyJhbGciOi...refresh.jwt
                summary: Refresh request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Refresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Refresh'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPairEnvelope'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                TokenInvalidOrExpired:
                  value:
                    error:
                      code: TOKEN_INVALID
                      message: Token is invalid or expired.
                  summary: Token invalid or expired
          description: ''
  /api/v1/auth/register/:
    post:
      operationId: auth_register
      description: Creates a new user and sends a verification email. Duplicate emails
        return 202 with a generic body to prevent enumeration.
      summary: Register a new account
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
            examples:
              RegisterRequest:
                value:
                  email: trader@example.com
                  display_name: Jane Trader
                  password: correct horse battery staple
                summary: Register request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Register'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterOkEnvelope'
          description: ''
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusEnvelope'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                PasswordDoesNotMeetPolicy:
                  value:
                    error:
                      code: PASSWORD_WEAK
                      message: Password does not meet policy.
                      details:
                        password:
                        - This password is too common.
                  summary: Password does not meet policy
                ValidationError:
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: Invalid input.
                      details:
                        email:
                        - Enter a valid email address.
                  summary: Validation error
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                RateLimited:
                  value:
                    error:
                      code: RATE_LIMITED
                      message: Too many requests.
                  summary: Rate limited
          description: ''
  /api/v1/auth/resend-verification/:
    post:
      operationId: auth_resend_verification
      description: Idempotent. Always returns 200 to avoid email enumeration.
      summary: Resend verification email
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerification'
            examples:
              ResendRequest:
                value:
                  email: trader@example.com
                summary: Resend request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ResendVerification'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResendVerification'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusEnvelope'
          description: ''
        '429':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                RateLimited:
                  value:
                    error:
                      code: RATE_LIMITED
                      message: Too many requests.
                  summary: Rate limited
          description: ''
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      required:
      - error
    Login:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          writeOnly: true
      required:
      - email
      - password
    Logout:
      type: object
      properties:
        refresh:
          type: string
      required:
      - refresh
    PasswordReset:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
      - email
    PasswordResetConfirm:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          writeOnly: true
          maxLength: 256
          minLength: 12
      required:
      - password
      - token
    Refresh:
      type: object
      properties:
        refresh:
          type: string
      required:
      - refresh
    Register:
      type: object
      properties:
        email:
          type: string
          format: email
        display_name:
          type: string
          maxLength: 64
          minLength: 1
        password:
          type: string
          writeOnly: true
          maxLength: 256
          minLength: 12
      required:
      - display_name
      - email
      - password
...
    RegisterOkEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RegisterData'
      required:
      - data
    ResendVerification:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
      - email
    StatusData:
      type: object
      properties:
        status:
          type: string
      required:
      - status
    StatusEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/StatusData'
      required:
      - data
...
    TokenPairEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/TokenPairData'
      required:
      - data
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT